+++ src/test/java/io/github/astrapi69/collection/list/ListExtensionsTest.java
@@ -804,6 +804,48 @@ public class ListExtensionsTest
 	}
 
 	/**
+	 * Test method for {@link ListExtensions#relocate(List, Object, int)}
+	 */
+	@Test
+	public void testRelocate()
+	{
+		List<String> list;
+		int index;
+
+		list = ListFactory.newArrayList();
+		list.add("Leonidas");
+		list.add("Berta");
+		list.add("Caesar");
+		list.add("Dora");
+		list.add("Emil");
+		list.add("Anton");
+
+		ListExtensions.relocate(list, "Anton", 0);
+		index = list.indexOf("Anton");
+		assertEquals(index, 0);
+
+		ListExtensions.relocate(list, "Anton", 2);
+		index = list.indexOf("Anton");
+		assertEquals(index, 2);
+
+		ListExtensions.relocate(list, "Anton", 5);
+		index = list.indexOf("Anton");
+		assertEquals(index, 5);
+
+		ListExtensions.relocate(list, "Anton", 6);
+		index = list.indexOf("Anton");
+		assertEquals(index, 5);
+
+		ListExtensions.relocate(list, "Anton", 8);
+		index = list.indexOf("Anton");
+		assertEquals(index, 5);
+
+		ListExtensions.relocate(list, "foo", 1);
+		index = list.indexOf("foo");
+		assertEquals(index, -1);
+	}
+
+	/**
 	 * Test method for {@link ListExtensions#removeFirstElement(List)}
 	 */
 	@Test
@@ -1238,7 +1280,7 @@ public class ListExtensionsTest
 		for (int i = 0; i < unsortedHelpMenu.size(); i++)
 		{
 			String currentUnsortedHelpMenuItem = unsortedHelpMenu.get(i);
-			int indexToInsert = ListExtensions.getIndexToInsert(helpMenu, newSortedHelpMenu,
+			int indexToInsert = ListExtensions.getIndexToInsert(newSortedHelpMenu, helpMenu,
 				currentUnsortedHelpMenuItem);
 			newSortedHelpMenu.add(indexToInsert, currentUnsortedHelpMenuItem);
 		}